fix(fuzz): fix fuzz test and refactor test utils
The fuzz test is now compatible with godwoken-v1. But we need to provide
more seeds to make it more effective.

The original rust test codes use `helper.rs` and there are a lot of
boilerplate codes to write a fresh new test.
A new `ctx.rs` is added to solve this problem. Considering moving all the
tests into new ones need a lot of effort, so the tests will be
refactored gradually.